+++ js/data.js
@@ -1,0 +1,26 @@
+import {getRandomPositiveInteger} from "./util.js"
+const ID = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25];
+const NAME = ['vitaly', 'volodya', 'kostya', 'nikita', 'sasha', 'maria'];
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const createDirections = () => ({
+  id: getRandomArrayElement(ID),
+  avatar: `img/avatar-${getRandomPositiveInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(MESSAGE),
+  name: getRandomArrayElement(NAME),
+  likes: getRandomPositiveInteger(15, 200),
+  url: `img/avatar-${getRandomArrayElement(ID)}.svg`
+});
+
+const directions = Array.from({length: 25}, createDirections);
+
+export {directions};
